Convert Hashcode to uint

+++ AElf.Kernel.Types/ChainHelpers.cs
@@ -16,7 +16,7 @@ namespace AElf.Common
             // For 4 base58 chars use following range (2111 ~ zzzz):
             // Max: 57*58*58*58+57*58*58+57*58+57 = 11316496 (zzzz)
             // Min: 1*58*58*58+0*58*58+0*58+0 = 195112 (2111)
-            var validNUmber = serialNumber.GetHashCode() % 11316496;
+            var validNUmber = (uint)serialNumber.GetHashCode() % 11316496;
             if (validNUmber < 195112)
                 validNUmber += 195112;
